MNT: move converter functions to dedicated module
