+++ src/databroker_elasticsearch/callback.py
@@ -5,51 +5,6 @@ from elasticsearch import Elasticsearch
 from elasticsearch import helpers as eshelpers
 
 
-def noconversion(x):
-    return x
-
-
-def toisoformat(epoch):
-    """Convert epoch seconds to elasticsearch friendly ISO time.
-
-    When `epoch` is a float return ISO date with millisecond
-    precision.  Otherwise return date rounded to seconds.
-
-    Parameters
-    ----------
-    epoch : float
-        The time in seconds since POSIX epoch in 1970.
-
-    Returns
-    -------
-    isodate : str
-        The ISO formatted date and time with second or millisecond precision.
-    """
-    from datetime import datetime
-
-    epochms = round(epoch, 3)
-    dt = datetime.fromtimestamp(epochms)
-    tiso = dt.isoformat()
-    rv = tiso[:-3] if dt.microsecond else tiso
-    assert len(rv) in (19, 23)
-    return rv
-
-
-def normalize_counts(d):
-    if not isinstance(d, dict):
-        return None
-    totalcount = sum(d.values()) or 1.0
-    rv = dict((k, v / totalcount) for k, v in d.items())
-    return rv
-
-
-def listofstrings(v):
-    rv = None
-    if isinstance(v, list) and all(isinstance(w, str) for w in v):
-        rv = v
-    return rv
-
-
 def esdocument(docmap, entry):
     rv = {}
     for doc_key, esname, fcnv in docmap:
